requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
PyJWT==2.6.0
python-i18n==0.3.9
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
